% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_face.R
\name{adjust_face}
\alias{adjust_face}
\title{adjust face}
\usage{
adjust_face(far, rar, u, v)
}
\arguments{
\item{far}{a face array data structure}

\item{rar}{a ridge array data structure}

\item{u}{in index of a face}

\item{v}{an index of an adjoining face}
}
\value{
a copy of the adjoining face which has been rotated, shifted,
and aligned so as to fit next to face u in the plane
}
\description{
adjust face
}
\examples{
far <- example_far
rar <- example_rar
adjust_face(far, rar, 1, 5)

}
