% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_face.R
\name{align_face}
\alias{align_face}
\title{Align face}
\usage{
align_face(face)
}
\arguments{
\item{face}{a face of a polyhedron, as a 'face_array' structure}
}
\value{
the face with its columns rearranged so as to pass the check_angles check
}
\description{
Align face
}
\examples{
face <- array(c(14,0,0,7,0,1,6,1,1,12,1,0), dim = c(3,4),
dimnames = list(c('nm', 'x', 'y'), letters[1:4]))
align_face(face)
}
