% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_face.R
\name{apply_difuv}
\alias{apply_difuv}
\title{apply_diff for u,v}
\usage{
apply_difuv(far, u, v)
}
\arguments{
\item{far}{a face array data structure}

\item{u}{in index of a face}

\item{v}{an index of an adjoining face}
}
\value{
a translation matrix which will shift the adjoining face
so that it aligns with cave u
}
\description{
apply_diff for u,v
}
\examples{
far <- example_far
apply_difuv(far, 1, 5)


}
