% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_face.R
\name{check_angles}
\alias{check_angles}
\title{Check angles}
\usage{
check_angles(angles)
}
\arguments{
\item{angles}{a vector of 4 angles, in degrees from 0 to 360}
}
\value{
TRUE if angles increase clockwise, else FALSE
}
\description{
Check angles
}
\examples{
check_angles(c(45, 135, 225, 315)) # expect TRUE
check_angles(c(145, 35, 225, 315)) # expect FALSE

}
