% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_d8.R
\docType{data}
\name{d8}
\alias{d8}
\title{d8}
\format{
An array with 8 rows and 4 variables:
\tabular{rrrrlll}{
\strong{w} \tab \strong{x} \tab \strong{y} \tab \strong{z} \tab \strong{op} \tab \strong{cycles} \tab \strong{description} \cr
1 \tab 2 \tab 3 \tab 4 \tab e   \tab ()       \tab noop                                                          \cr
3 \tab 4 \tab 1 \tab 2 \tab a2  \tab (13)(24) \tab rotation by 180                                               \cr
4 \tab 1 \tab 2 \tab 3 \tab a3  \tab (1432)   \tab rotation by 270 counter-clockwise                             \cr
2 \tab 3 \tab 4 \tab 1 \tab a   \tab (1234)   \tab rotation by 90 counter-clockwise                              \cr
1 \tab 4 \tab 3 \tab 2 \tab a2x \tab (24)     \tab reflection about the diagonal 1-3                             \cr
3 \tab 2 \tab 1 \tab 4 \tab x   \tab (13)     \tab reflection about the diagonal 2-4                             \cr
4 \tab 3 \tab 2 \tab 1 \tab ax  \tab (14)(23) \tab reflection about the line joining midpoints of sides 14 and 23\cr
2 \tab 1 \tab 4 \tab 3 \tab a3x \tab (12)(34) \tab reflection about the line joining midpoints of sides 12 and 34
}
}
\source{
SageMath 9.5:

\code{G = DihedralGroup(4)
print(gap.StructureDescription(G), gap.IdGroup(G))
for per in G:
  d = per.dict()
  print(tuple(d.values()))
}
}
\usage{
d8
}
\description{
Description. (from \url{https://groupprops.subwiki.org/wiki/Dihedral_group:D8#Geometric_definition})
}
\details{
The dihedral group D_8 (also called D_4) is defined as
the group of all symmetries of the square (the regular 4-gon).

This has a cyclic subgroup comprising rotations
(which is the cyclic subgroup generated by a)
and has four reflections each being an involution:
reflections about lines joining midpoints of opposite sides,
and reflections about diagonals.
}
\examples{

e   <- layoutsquares::d8[1,]
a2  <- layoutsquares::d8[2,]
a3  <- layoutsquares::d8[3,]
a   <- layoutsquares::d8[4,]
a2x <- layoutsquares::d8[5,]
x   <- layoutsquares::d8[6,]
ax  <- layoutsquares::d8[7,]
a3x <- layoutsquares::d8[8,]

d8ops <- list(e, a, a2, a3, x, ax, a2x, a3x)

# https://groupprops.subwiki.org/wiki/Dihedral_group:D8#Multiplication_table
multiplication_table <- list(
  list(e, a, a2, a3, x, ax, a2x, a3x),
  list(a, a2, a3, e, ax, a2x, a3x, x),
  list(a2, a3, e, a, a2x, a3x, x, ax),
  list(a3, e, a, a2, a3x, x, ax, a2x),
  list(x, a3x, a2x, ax, e, a3, a2, a),
  list(ax, x, a3x, a2x, a, e, a3, a2),
  list(a2x, ax, x, a3x, a2, a, e, a3),
  list(a3x, a2x, ax, x, a3, a2, a, e)
)

for(i in 1:8) {
  for(j in 1:8) {
    opi <- d8ops[[i]]
    opj <- d8ops[[j]]
    prod <- multiplication_table[[i]][[j]]
    print(sprintf("\%d,\%d: c(\%s)[c(\%s)] ==  c(\%s)", i,j,
      stringr::str_c(opi, collapse=','),
      stringr::str_c(opj, collapse=','),
      stringr::str_c(prod, collapse=',')))
    stopifnot(all(opi[opj] == prod))
  }
}

rm(e, a, a2, a3, x, ax, a2x, a3x, d8ops)

}
\keyword{datasets}
