% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_first.R
\name{depth_first_gr}
\alias{depth_first_gr}
\title{Depth First}
\usage{
depth_first_gr(g)
}
\arguments{
\item{g}{an igraph graph}
}
\value{
a list of indices of edges that are traversed in depth first order
}
\description{
Recurse through the adjacency matrix of the graph, looking for row/column
intersections, while keeping track of intersections that have already been seen
}
\examples{
graph <- igraph_from_text("KP@?Y_AG?C?G")[[1]]
depth_first_gr(graph)

}
