% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_example_far.R
\docType{data}
\name{example_far}
\alias{example_far}
\title{example_far}
\format{
A data frame with 3 rows and 4 variables:
\describe{
}
}
\source{
Source
}
\usage{
example_far
}
\description{
Description.
}
\keyword{datasets}
