% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_example_rar.R
\docType{data}
\name{example_rar}
\alias{example_rar}
\title{example_rar}
\format{
A data frame with 24 rows and 4 variables:
\describe{
}
}
\source{
Source
}
\usage{
example_rar
}
\description{
Description.
}
\keyword{datasets}
