% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_face.R
\name{get_angle}
\alias{get_angle}
\title{Get angle}
\usage{
get_angle(cent, perif)
}
\arguments{
\item{cent}{the center point}

\item{perif}{the periferal point}
}
\value{
a number between 0 and 360
}
\description{
Given a center point and a periferal point, determine the angle
of the corresponding plane vector with respect to the line pointing
towards the 'South Pole', going clockwise from 6:00
}
\examples{
get_angle(c(0,0), c(1,1)) # expect 225

}
