% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_face.R
\name{get_angles}
\alias{get_angles}
\title{get_angles}
\usage{
get_angles(mx)
}
\arguments{
\item{mx}{a 2x4 matrix of vertex coordinates}
}
\value{
a vector of angles for the vertices in mx
}
\description{
get_angles
}
\examples{
face <- array(c(7,0,0,4,0,1,10,1,1,14,1,0), dim = c(3,4),
dimnames = list(c('nm', 'x', 'y'), letters[1:4]))
mx <- face[c('x','y'),]
get_angles(mx) # expect 45 135 225 315

}
