% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_face.R
\name{get_opuv}
\alias{get_opuv}
\title{get_op for u, v}
\usage{
get_opuv(far, rar, u, v)
}
\arguments{
\item{far}{a face array data structure}

\item{rar}{a ridge array data structure}

\item{u}{in index of a face}

\item{v}{an index of an adjoining face}
}
\value{
the operator which, when applied to face v, aligns v with u
}
\description{
get_op for u, v
}
\examples{
far <- example_far
rar <- example_rar
get_opuv(far, rar, 1, 3)

}
