% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_face.R
\name{get_other_edge}
\alias{get_other_edge}
\title{Get other edge}
\usage{
get_other_edge(edge)
}
\arguments{
\item{edge}{an edge of a face}
}
\value{
the adjoining edge of the adjoining face
}
\description{
Get other edge
}
\examples{
xx <- c(1,4)
yy <- c(2,3)
identical(get_other_edge(xx), yy)
}
