% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutsquares-package.R
\docType{package}
\name{layoutsquares-package}
\alias{layoutsquares}
\alias{layoutsquares-package}
\title{layoutsquares: Layout Squares on a Plane}
\description{
Layout the faces of a polyhedron as squares in a plane.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jefflansing.github.io/layoutsquares/}
}

}
\author{
\strong{Maintainer}: Jeff Lansing \email{lansing2357@gmail.com} (\href{https://orcid.org/0009-0001-2532-9625}{ORCID})

}
\keyword{internal}
