% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_face.R
\docType{data}
\name{nm}
\alias{nm}
\title{nm}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
nm
}
\value{
the index of the row in a face_array data structure that
contain the names of the corners of the faces
}
\description{
nm
}
\examples{
far <- example_far
nm <- 1
far[nm,,1]

}
\keyword{datasets}
