% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_oar.R
\docType{data}
\name{oar}
\alias{oar}
\title{oar}
\format{
A data frame with 4 rows and 4 variables:
\describe{\tabular{cccc}{
   a \tab b \tab c \tab d \cr
   1 \tab 2 \tab 4 \tab 3 \cr
   2 \tab 3 \tab 1 \tab 4 \cr
   3 \tab 4 \tab 2 \tab 1 \cr
   4 \tab 1 \tab 3 \tab 2 \cr
   } \tab  \tab  \tab  \cr
}
}
\source{
Source
}
\usage{
data(oar)
}
\description{
Description:
A look-up table based on tiling the plane with oriented squares,
}
\keyword{datasets}
