% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_face.R
\name{selectby_uv}
\alias{selectby_uv}
\title{select by u and v}
\usage{
selectby_uv(rar, u, v)
}
\arguments{
\item{rar}{a ridge array data structure}

\item{u}{an index of a face in r}

\item{v}{an index of another face in rar}
}
\value{
the row in rar that contains both faces
}
\description{
SELECT * FROM rar WHERE (f1 = u AND f2 = v) OR (f1 = v AND f2 = u)
}
\examples{
rar <- example_rar
selectby_uv(rar, 1,5)[3:4]

}
