% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_face.R
\docType{data}
\name{xy}
\alias{xy}
\title{xy}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
xy
}
\value{
the indices of the rows in a face_array data structure that
contain the x and y coordinates
}
\description{
xy
}
\examples{
far <- example_far
xy <- c(2,3)
far[xy,,1]

}
\keyword{datasets}
