% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_first.R
\name{get_evids}
\alias{get_evids}
\title{Get edge vertex ids}
\usage{
get_evids(gr, vnames)
}
\arguments{
\item{gr}{an igraph graph object}

\item{vnames}{a character vector of vertex names}
}
\value{
a matrix of vertix ids
}
\description{
First call depth_first_gr, then convert the edge ids to pairs of vertex ids,
using a vector of vertex names for guidance
}
\examples{
vnames <- as.character(1:14)
graph <- rgraph6::igraph_from_text("KP@?Y_AG?C?G")[[1]]
get_evids(graph, vnames)

}
