% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_oar.R
\docType{data}
\name{oar}
\alias{oar}
\title{oar}
\format{
An array with 4 rows and 4 variables:
\tabular{rrrr}{
\strong{a} \tab \strong{b} \tab \strong{c} \tab \strong{d} \cr
1 \tab 2 \tab 4 \tab 3\cr
2 \tab 3 \tab 1 \tab 4\cr
3 \tab 4 \tab 2 \tab 1\cr
4 \tab 1 \tab 3 \tab 2
}
}
\source{
Source
}
\usage{
oar
}
\description{
Description.
}
\keyword{datasets}
